Move Title component to typography folder
